refactor curseforge download resolution
